requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0
django-allauth==0.40.0
django-braces==1.13.0
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
multidict==4.7.1
oauthlib==3.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.19
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==5.0.1
yarl==1.4.2
